+++ src/components/common/back-link-bar.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import styles from '@/styles/components/common/back-link.module.scss';
+import styles from '@/styles/components/common/links/back-link.module.scss';
 import ArrowCircleLeftIcon from '@mui/icons-material/ArrowCircleLeft';
 import { PagePathname } from '@/src/constants/strings';