+++ src/components/common/buttons/button.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Link from 'next/link';
-import styles from '@/styles/components/common/button.module.scss';
+import styles from '@/styles/components/common/buttons/button.module.scss';
 
 import LoginIcon from '@mui/icons-material/Login';
 import LogoutIcon from '@mui/icons-material/Logout';