+++ src/components/common/links/edit-link.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import EditIcon from '@mui/icons-material/Edit';
-import styles from '@/styles/components/common/add-button.module.scss';
+import styles from '@/styles/components/common/buttons/add-button.module.scss';
 import { useRouter } from 'next/navigation';
 import { PagePathname } from '@/src/constants/strings';